Phonebook for school project (C++)
This is an examination project that I did during my first year in Computer Engineering. We had two opions; 'Bulls and Cows' or 'Phonebook'. I chose the second alternative, Phonebook, as I had previously been overwhelmed by looking at my classmate's code for Bulls and Cows. Since I already had a decent amount of experience with in-and-out textfile-streaming I felt it was best to stick to my comfort zone.

My final grade for the project was A. My code consists of error handling and crash-preventing statements/algorithms.

If you have any suggestions of improvements, or questions, please feel free to comment or add me on Discord (Milton#8481).